TCPL-CH01: Fixed some long and int type mismatch

+++ the_c_prog_lang/ch01/longest_line.c
@@ -2,7 +2,7 @@
 
 #define MAX_LINE_LENGTH 1000
 
-long get_line(char *line) {
+int get_line(char *line) {
     int i = 0;
     int c = 0;
     while((c = getchar()) != EOF) {
@@ -31,15 +31,15 @@ void copy_from_to(char * from, char * to) {
 int main() {
     char max_line[MAX_LINE_LENGTH];
     char line[MAX_LINE_LENGTH];
-    long max_line_length = 0;
-    long line_length = 0;
+    int max_line_length = 0;
+    int line_length = 0;
     while ((line_length = get_line(line)) > 0) {
-        // printf("Line = %s, length = %ld\n", line, line_length);
+        // printf("Line = %s, length = %d\n", line, line_length);
         if (line_length > max_line_length) {
             copy_from_to(line, max_line);
             max_line_length = line_length;
         }
     }
-    printf("Max line = %s, length = %ld\n", max_line, max_line_length);
+    printf("Max line = %s, length = %d\n", max_line, max_line_length);
     return 0;
 }
